move core of the checking to a task
